separate Mutation and Condition
